Created new library "mcutils", cleaned Minizinc
Old library "mcputils.py" is still available but should be considered deprecated, please switch to the other and integrate it with anything deemed necessary. Check "Minizinc/Minizinc.py" to see how to import it.

MILP file is only a copy-paste of the ipynb for now.

Created all 21 JSONs for Minizinc.